fix(Submission Form): :bug: Fixed issue with stale form state
For whatever reason, I was both setting the value for hasOpenCall AND resetting the form with the values to include it immediately after. Fixed that and resolved the issues with the org name not resetting, the dates, the open call presence check, etc. Basically everything. At any rate, everything should be cleaned up. Also adjusted the open-call-card to utilize event.type and not eventType. Hopefully the last remnant"