refactor(Auth): :memo: Sign up/sign in updates
Added proper loading state and button disables for success, pending, and loading states (respectively) and adjusted the register card to autohandle the otp when the length is correct and the value is different from previous attempts. Also adjusted navbars to just check for user rather than using Unauthenticated as it was causing lags between the two.