fix(Submission Form): :zap: Filepond and file previewer updates. Submission form tempFiles save handling update
Updated the filepond definitions to include excel files as well as webp. Still didn't set file sizes for either. Also updated the file previewer to accept files that use the office file types and updated the submission forms to autosave (uploading the files) when new files are selected and valid.